feat: package structure reorganization
