+++ app/src/main/java/com/example/emergetestapplication/emerge/presentation/view/compose/SignUpScreen.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.emergetestapplication.R
+import com.example.emergetestapplication.emerge.presentation.view.state.AuthState
 import com.example.emergetestapplication.ui.theme.EmergeTestApplicationTheme
 
 @Composable
